Revert "Fix compilation warnings"
This reverts commit 08fbef11ab1250b61b5f014010a9d16e0e1e6aaf.

+++ reaper_csurf_integrator/control_surface_integrator.cpp
@@ -3354,7 +3354,7 @@ void TrackNavigationManager::RebuildTracks()
             for (; trackID <= GetNumTracks(); trackID++)
             {
                 MediaTrack* track = CSurf_TrackFromID(trackID, followMCP_);
-                int depthOffset = static_cast<int>(GetMediaTrackInfo_Value(track, "I_FOLDERDEPTH"));
+                int depthOffset = GetMediaTrackInfo_Value(track, "I_FOLDERDEPTH");
 
                 if (trackID == currentFolderTrackID_)
                 {
